+++ app/src/main/java/pt/ipleiria/mytodo/models/Group.kt
@@ -1,9 +1,11 @@
 package pt.ipleiria.mytodo.models
 
+import java.util.*
+
 data class Group(
         override var id: String = "",
         var owner: String = "",
         var name: String = "",
         var members: List<String> = listOf(),
-        var todos: List<Todo> = listOf()
+        override var timestamp: Date? = Date()
 ): Base