+++ app/src/main/java/pt/ipleiria/mytodo/models/Todo.kt
@@ -5,6 +5,6 @@ import java.util.*
 data class Todo(
     override var id: String = "",
     var text: String = "",
-    var date: Date = Date(),
-    var by: String= ""
+    var by: String= "",
+    override var timestamp: Date? = Date(),
 ): Base