+++ app/src/main/java/pt/ipleiria/mytodo/ui/todos/viewModels/TodoFormState.kt
@@ -1,0 +1,9 @@
+package pt.ipleiria.mytodo.ui.todos.viewModels
+
+/**
+ * Data validation state of the login form.
+ */
+data class TodoFormState(
+    val textError: Int? = null,
+    val isDataValid: Boolean = false
+)